Reorganize Controllers for Admin and USers
