+++ src/Controller/OrphanController.php
@@ -43,53 +43,4 @@ class OrphanController extends AbstractController
             'form' => $form->createView(),
         ]);
     }
-
-    /**
-     * @Route("/admin/orphan/{id}")
-     * @IsGranted("ROLE_ADMIN")
-     */
-    public function show(OrphanUser $orphanUser): Response
-    {
-        return $this->render('orphan_user/show.html.twig', [
-            'orphan_user' => $orphanUser,
-        ]);
-    }
-
-    /**
-     * @Route("/admin/orphan/edit/{id}")
-     * @IsGranted("ROLE_ADMIN")
-     */
-    public function edit(OrphanUser $orphanUser, Request $request): Response
-    {
-        $form = $this->createForm(OrphanUserType::class, $orphanUser);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
-
-            return $this->redirectToRoute('app_orphan_edit', [
-                'orphan' => $orphanUser->getId(),
-            ]);
-        }
-
-        return $this->render('orphan_user/edit.html.twig', [
-            'orphan_user' => $orphanUser,
-            'form' => $form->createView(),
-        ]);
-    }
-
-    /**
-     * @Route("/admin/orphan/delete/{id}")
-     * @IsGranted("ROLE_ADMIN")
-     */
-    public function delete(Request $request, OrphanUser $orphanUser): Response
-    {
-        if ($this->isCsrfTokenValid('delete'.$orphanUser->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($orphanUser);
-            $entityManager->flush();
-        }
-
-        return $this->redirectToRoute('home');
-    }
 }
