+++ src/Controller/ProductController.php
@@ -20,65 +20,20 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class ProductController extends AbstractController
 {
-
-    public function create(Request $request, FileService $fileService, ProductService $productService): Response
+    public function show(Product $product, ProductService $productService): Response
     {
-        $product = new Product();
-
-        $form = $this->createForm(ProductType::class, $product);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-
-            foreach ($product->getFiles() as $file) {
-
-                $file->setProduct($product);
-                $fileService->saveFile($file);
-            }
-
-            $productService->newProduct($product);
-
-            return $this->redirectToRoute('home');
-        }
-
-        return $this->render('product/new.html.twig', [
-            'product' => $product,
-            'form' => $form->createView(),
-        ]);
-    }
-
-    public function show(Product $product): Response
-    {
-        if (!$this->getUser() === $product->getUser() & !$this->isGranted("ROLE_ADMIN")){
-            $this->addFlash("warning", "Vous n'avez pas accès à cet item.");
-
-            return $this->redirectToRoute('home');
-        }
+        if(!$productService->checkPermission($product)) return $this->redirectToRoute('home');
 
         return $this->render('product/show.html.twig', [
             'product' => $product,
         ]);
     }
 
-    public function myProducts()
+    public function edit(Request $request, Product $product, FileService $fileService, ProductService $productService)
     {
-        $user = $this->getUser();
+        if(!$productService->checkEditable($product)) return $this->render("product/show.html.twig", ['product' => $product]);
 
-        return $this->render("product/index.html.twig", [
-            'products' => $user->getProducts()
-        ]);
-
-    }
-
-    public function edit(Request $request, Product $product, FileService $fileService)
-    {
-        if ( $product->getState() != 'En attente' and !$this->isGranted("ROLE_ADMIN")) {
-            $this->addFlash("warning", "Vous ne pouvez pas modifier une demande une fois validée, contactez Lucille !");
-
-            return $this->redirectToRoute("app_userspace_showproduct", [
-                'product' => $product->getId()
-            ]);
-        }
+        if(!$productService->checkPermission($product)) return $this->redirectToRoute('home');
 
         $form = $this->createForm(ProductType::class, $product);
         $form->handleRequest($request);
@@ -114,15 +69,9 @@ class ProductController extends AbstractController
         ]);
     }
 
-    public function delete(Request $request, Product $product): Response
+    public function delete(Request $request, Product $product, ProductService $productService): Response
     {
-        if ($product->getState() != 'En attente' && !$this->isGranted("ROLE_ADMIN")) {
-            $this->addFlash("warning", "Vous ne pouvez pas modifier une demande une fois validée, contactez Lucille !");
-
-            return $this->render("product/show.html.twig", [
-                'product' => $product
-            ]);
-        }
+        if(!$productService->checkEditable($product)) return $this->render("product/show.html.twig", ['product' => $product]);
 
         if ($this->isCsrfTokenValid('delete'.$product->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
@@ -133,45 +82,5 @@ class ProductController extends AbstractController
         }
 
         return $this->redirectToRoute('home');
-    }
-
-    public function uploadBill(Request $request, Product $product, FileService $fileService)
-    {
-        $bill = new Bill();
-
-        $form = $this->createForm(BillType::class, $bill);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $product->setBill($form->getData());
-            $fileService->saveBill($form->getData());
-
-            return $this->redirect($request->request->get('referer'));
-        }
-
-        return $this->render('bill&devis/index.html.twig', [
-            'product' => $product,
-            'form' => $form->createView(),
-        ]);
-    }
-
-    public function uploadDevis(Request $request, Product $product, FileService $fileService)
-    {
-        $devis = new Devis();
-
-        $form = $this->createForm(DevisType::class, $devis);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $product->setDevis($form->getData());
-            $fileService->saveDevis($form->getData());
-
-            return $this->redirect($request->request->get('referer'));
-        }
-
-        return $this->render('bill&devis/index.html.twig', [
-            'product' => $product,
-            'form' => $form->createView(),
-        ]);
     }
 }
