+++ src/Controller/UserSpaceController.php
@@ -7,6 +7,7 @@ use App\Entity\Devis;
 use App\Entity\File;
 use App\Entity\Product;
 use App\Entity\User;
+use App\Form\ProductType;
 use App\Form\UserType;
 use App\Service\FileService;
 use App\Service\ProductService;
@@ -51,17 +52,19 @@ class UserSpaceController extends AbstractController
      */
     public function showOwnedProducts()
     {
-        return $this->render('product/index.html.twig', [
-            'products' => $this->getUser()->getProducts()
+        $user = $this->getUser();
+
+        return $this->render("product/index.html.twig", [
+            'products' => $user->getProducts()
         ]);
     }
 
     /**
      * @Route("/product-{product}")
      */
-    public function showProduct(Product $product, ProductController $controller)
+    public function showProduct(Product $product, ProductController $controller, ProductService $productService)
     {
-        return $controller->show($product);
+        return $controller->show($product, $productService);
     }
 
     /**
@@ -69,23 +72,44 @@ class UserSpaceController extends AbstractController
      */
     public function createProduct(Request $request, ProductController $controller, FileService $fileService, ProductService $productService)
     {
-        return $controller->create($request, $fileService, $productService);
+        $product = new Product();
+
+        $form = $this->createForm(ProductType::class, $product);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+
+            foreach ($product->getFiles() as $file) {
+
+                $file->setProduct($product);
+                $fileService->saveFile($file);
+            }
+
+            $productService->newProduct($product);
+
+            return $this->redirectToRoute('home');
+        }
+
+        return $this->render('product/new.html.twig', [
+            'product' => $product,
+            'form' => $form->createView(),
+        ]);
     }
 
     /**
      * @Route("/product/edit-{product}")
      */
-    public function editProduct(Product $product, ProductController $controller, Request $request, FileService $fileService)
+    public function editProduct(Product $product, ProductController $controller, Request $request, FileService $fileService, ProductService $productService)
     {
-        return $controller->edit($request,$product, $fileService);
+        return $controller->edit($request,$product, $fileService, $productService);
     }
 
     /**
      * @Route("/product/delete-{product}")
      */
-    public function deleteProduct(Product $product, ProductController $controller, Request $request)
+    public function deleteProduct(Product $product, ProductController $controller, Request $request, ProductService $productService)
     {
-        return $controller->delete($request,$product);
+        return $controller->delete($request,$product, $productService);
     }
 
     /**
